requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
contextlib2==21.6.0
cryptography==43.0.1
Django==5.1.1
django-ninja==1.3.0
django-ninja-extra==0.21.4
django-ninja-jwt==5.3.3
fire==0.6.0
injector==0.22.0
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2==2.9.9
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
rav==0.0.9
rich==13.8.1
six==1.16.0
sqlparse==0.5.1
termcolor==2.4.0
typing_extensions==4.12.2
